requirements.txt
pytz==2018.9
Pillow==5.4.1
Werkzeug==0.14.1